Friendly slugs and URLs for exercises
We want unique slugs for exercise Git URLs. As long as we're adding them
for Git URLs, we may as well use them for HTML URLs as well.

This changes the URLs from:

    /exercises/5

To:

    /exercises/null-object-part-two

This uses the friendly_id gem to generate and store the slugs as well
as to handle `to_param` and `find`.

https://trello.com/c/rfQ5yueH/2-as-an-admin-when-i-create-an-exercise-in-whetstone-it-creates-a-git-repository-for-the-exercise