Users are now guided through a review workflow
* Status messages are now shown in the header, along with calls-to-action like
"Continue" or "Next exercise"
* A user moves through Submitted, ReviewingSolution, AwaitingReview, and
AllStepsCompleted states.
* This commit introduces a `Status` namespace.
* `Status::Finder` looks for the appropriate Status to show, choosing
the first `Status::X` class that returns `true` for a `#applicable?`
method. This class' `to_partial_path` method is called to render the
appropriate partial.
* Commits including styles and markup for an unimplemented progress
bar were pushed to this branch, so one call to `render 'progress_bar'`
is commented out.

Possibly surprising:

This commit ignores whether user has commented on assigned solution.
Instead, we look at whether they've commented on _any_ solution.

This is necessary since a user's assigned solution changes as soon as
they review a solution. I'd call this a bug, or at least poor UX (the
"review this solution" prompt switches solutions and never disappears),
but fixing it is out of the scope of this feature.

I've created
https://trello.com/c/bHjpOxyb/109 to track this additional work.

http://trello.com/c/y9hKVEFb/78