Replace N+1 query with query object
Inline comments would be queried for each line in a diff during
rendering. This caused rendering diffs to be very slow.

This commit introduces a `InlineCommentQuery` object. Inline comments
are queried once for the entire revision and passed to that object which
then filters them using `Enumerable#select` when rendering each line,
saving a trip to the database.

This solution cut page load time by a factor of 2 on my local machine.