Show the SSH fingerprint
When the user clones the repo over SSH, they will be prompted to trust
the server and shown the fingerprint. Show the fingerprint on the Web
page in the clone instructions so they can make this decision with
knowledge.

This adds an environment variable, `GIT_SERVER_FINGERPRINT`, to be set
to the git server's SSH fingerprint. One can look at a server's
fingerprint using ssh-keygen(1):

    ssh-keygen -l -f /etc/ssh/ssh_host_ecdsa_key.pub